Struct w5500::W5500

pub struct W5500<'a> { /* fields omitted */ }

Methods

impl<'a> W5500<'a>

pub fn new<E, S: FullDuplex<u8, Error = E>>(
spi: &mut S,
cs: &'a mut dyn OutputPin
) -> W5500<'a>

pub fn init<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>
) -> Result<(), E>

pub fn reset<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>
) -> Result<(), E>

pub fn set_mode<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
wol: bool,
ping_block: bool,
ppoe: bool,
force_arp: bool
) -> Result<(), E>

pub fn set_interrupt_mask<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
sockets: &[Socket]
) -> Result<(), E>

pub fn set_socket_interrupt_mask<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
interrupts: &[Interrupt]
) -> Result<(), E>

pub fn set_gateway<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
gateway: &IpAddress
) -> Result<(), E>

pub fn set_subnet<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
subnet: &IpAddress
) -> Result<(), E>

pub fn set_mac<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
mac: &MacAddress
) -> Result<(), E>

pub fn get_mac<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>
) -> Result<MacAddress, E>

pub fn set_ip<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
ip: &IpAddress
) -> Result<(), E>

pub fn is_interrupt_set<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
interrupt: Interrupt
) -> Result<bool, E>

pub fn reset_interrupt<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
interrupt: Interrupt
) -> Result<(), E>

pub fn send_udp<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
local_port: u16,
host: &IpAddress,
host_port: u16,
data: &[u8]
) -> Result<(), E>

pub fn listen_udp<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
port: u16
) -> Result<(), E>

pub fn try_receive_udp<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
socket: Socket,
destination: &mut [u8]
) -> Result<Option<(IpAddress, u16, usize)>, E>

TODO destination buffer has to be as large as the receive buffer or complete read is not guaranteed

pub fn read_u8<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register
) -> Result<u8, E>

pub fn read_u16<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register
) -> Result<u16, E>

pub fn read_ip<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register
) -> Result<IpAddress, E>

pub fn read_from<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register,
target: &mut [u8]
) -> Result<(), E>

pub fn write_u8<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register,
value: u8
) -> Result<(), E>

pub fn write_u16<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register,
value: u16
) -> Result<(), E>

pub fn write_to<E>(
&mut self,
spi: &mut dyn FullDuplex<u8, Error = E>,
register: Register,
data: &[u8]
) -> Result<(), E>